Remove unlock_pool_shared function
It is not shared any more.
